Brain dump commit.
Sprinkled in some new references to the introduction and
added comments with content I'd like to incorporate into the
intro. Also changed the mesh in `mesh_deformation.pdf` and
added a secondary figure with that mesh plus an extra layer
on the outside.

Also adding BLUE, GREEN and RED to `plot_utils`.